Initial, crude version.
This initial, pre-alpha version of the 'Verse contains all bodies and orbits as listed in "The Verse in Numbers", version 2.1, except for asteroid belts.  Planets and moons are rendered crudely: terraformed bodies are green, those in the process of being terraformed are yellow, others are red.  Gas giants are rendered with generic gas-giant textures that come with Celestia.  Much work remains to be done.